SQL requests were added

+++ NTI_APP/NTI_APP/Program.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Diagnostics;
+using MySql.Data.MySqlClient;
 
 namespace NTI_APP
 {
@@ -15,6 +16,17 @@ namespace NTI_APP
         [STAThread]
         static void Main()
         {
+            var cb = new MySqlConnectionStringBuilder();
+            cb.Server = "localhost";
+            cb.UserID = "root";
+            cb.Password = "123";
+            cb.Database = "baumancitydb";
+            var conn = new MySqlConnection(cb.ConnectionString);
+            //string connStr = "server=localhost;user=root;database=baumancitydb;password=123;";
+            //MySqlConnection conn = new MySqlConnection(connStr);
+
+            conn.Open();
+
             ProcessStartInfo procInfo = new ProcessStartInfo(@"C:/Program Files (x86)/Real Games/Factory IO/Factory IO.exe");
             procInfo.UseShellExecute = true;
             procInfo.FileName = "C:/Program Files (x86)/Real Games/Factory IO/Scenes/Automated Warehouse.factoryio";
@@ -34,10 +46,13 @@ namespace NTI_APP
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new MainForm(factoryIOProcess));
+            Application.Run(new MainForm(factoryIOProcess, conn));
 
-            factoryIOProcess.Kill();
-            factoryIOProcess.WaitForExit();
+            if (factoryIOProcess != null)
+            { 
+                factoryIOProcess.Kill();
+                factoryIOProcess.WaitForExit();
+            }
         }
     }
 }
